requirements.txt
absl-py==2.3.0
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cloudinary==1.42.2
cryptography==44.0.2
daraja-mpesa==1.2.1
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-daraja==1.3.0
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
dotenv==0.9.9
future==1.0.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
immutabledict==4.2.1
numpy==2.2.6
oauthlib==3.2.2
ortools==9.12.4544
pandas==2.2.3
protobuf==5.29.5
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-daraja==1.2.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
schema==0.7.7
six==1.17.0
sqlparse==0.5.3
tenacity==9.1.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uuid==1.30
